extract flask app logic
