+++ Procfile
@@ -1,2 +1,2 @@
 # web: python3 pt_bot.py
-web: gunicorn -c gun_config.py notifier_web:app
+web: gunicorn -c gun_config.py app:app
